% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atl_data.R
\docType{data}
\name{atl_data}
\alias{atl_data}
\alias{atl_env}
\alias{atl_spp}
\title{Atlantic Planktic Foraminifera}
\format{
An object of class \code{data.frame} with 61 rows and 3 columns.

An object of class \code{data.frame} with 61 rows and 22 columns.
}
\usage{
atl_env

atl_spp
}
\description{
Planktic foramifera relative abundances and environmental data in 61 coretop samples from the Atlantic Ocean
}
\details{
Datasets atl_spp and atl_env are data.frames containing foraminifera species and environmental data on 61 coretops from the Atlantic Ocean.
}
\examples{
data(atl_env)
data(Atl_spp)
}
\references{
Imbrie and Kipp (1971)
}
\keyword{datasets}
