% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.plot.r
\name{plot_grid}
\alias{plot_grid}
\title{plot grid}
\usage{
plot_grid(x, choices = 1:2)
}
\arguments{
\item{x}{ordination result from which site scores can be extracted or data.frame of values to plot}

\item{choices}{which ordination axes to plot, vector of two interger.}
}
\description{
Plots the grid of sites in the simulated data
}
\details{
Sites are shown by dots, lines show which sites have the same value for environmental variables 1 and 2 in the simulated data.
}
\note{
This function will be of little or no use outside of Bio303 indirect ordination practical as sites are rarely on a perfect grid.
}
\examples{
data(sim_env2)
plot_grid(sim_env2)

library(vegan)
DCA <- decorana(sim_spp2)
plot_grid(DCA, choices = 1:2)

}
