% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponds_data.R
\docType{data}
\name{ponds_data}
\alias{ponds_data}
\alias{ponds_env}
\alias{ponds_spp}
\title{Ponds data}
\format{
An object of class \code{data.frame} with 30 rows and 15 columns.

An object of class \code{data.frame} with 30 rows and 48 columns.
}
\usage{
ponds_env

ponds_spp
}
\description{
Diatom relative abundances and environmental data in 30 English ponds
}
\details{
Datasets pond_spp and pond_env are data.frames containing diatoms species and environmental data on 30 English ponds.
}
\examples{
data(ponds_env)
data(ponds_spp)
}
\references{
Bennion, H. 1994. A diatom-phosphorus transfer function for shallow, eutrophic ponds in southeast England. \emph{Hydrobiologia} 275-76: 391-410
}
\keyword{datasets}
