% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\alias{sim_env1}
\alias{sim_env2}
\alias{sim_spp1}
\alias{sim_spp2}
\title{Simulated environmental and species data}
\format{
An object of class \code{data.frame} with 100 rows and 2 columns.

An object of class \code{data.frame} with 100 rows and 2 columns.

An object of class \code{data.frame} with 100 rows and 27 columns.

An object of class \code{data.frame} with 100 rows and 21 columns.
}
\usage{
sim_env1

sim_env2

sim_spp1

sim_spp2
}
\description{
Simulate environmental and species data to test how well ordination methods can capture the underlying structure in species data.
}
\details{
Datasets sim_env1 and sim_env2 contain environmental data simulated on a grid. In the former case, the grid is square, in the latter, one environmental variable is more important than the other.

Datasets sim_spp1 and sim_spp2 contain simulated species data for the environmental data.
}
\examples{
data(sim_spp1)
library(vegan)
decorana(sim_spp1)
  
}
\keyword{datasets}
